server/prefect/requirements.txt
absl-py==0.1.6
appdirs==1.4.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
croniter==0.3.37
dask==2021.3.0
distributed==2021.3.0
docker==4.4.4
HeapDict==1.0.1
idna==2.10
immutables==0.14
importlib-metadata==2.1.1
marshmallow==3.10.0
marshmallow-oneofschema==2.1.0
msgpack==1.0.2
mypy-extensions==0.4.3
natsort==7.1.1
pendulum==2.1.2
prefect==0.14.12
psutil==5.8.0
psycopg2-binary==2.8.6
pyee==8.1.0
pyppeteer==0.2.5
python-box==5.3.0
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sodapy==2.1.0
sortedcontainers==2.3.0
tabulate==0.8.9
tblib==1.7.0
text-unidecode==1.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.59.0
urllib3==1.26.4
websocket-client==0.58.0
websockets==8.1
zict==2.0.0
zipp==3.4.1
